// src/utils/types.ts
export interface CommonClassesInterface {
  labelClass?: string;
  errorClass?: string;
  fieldClass?: string;
  wrapperClass?: string;
  optionLabelClass?: string;
}
// ...
export type FundingFormFieldType =
  | RangeControllerProps
  | InputControllerProps
  | DropdownControllerProps
  | DateControllerProps
  | TextAreaControllerProps
  | fileControllerProps
  | CheckBoxControllerProps
  | RadioButtonControllerProps
  | MultiCheckBoxControllerProps;

export interface RangeControllerProps extends Partial<CommonClassesInterface> {
  type: 'range';
  key?: string;
  name: string;
  label?: string;
  placeholder?: string;
  isRequired?: boolean;
  defaultValue?: string;
  isDisabled?: boolean;
  min?: number;
  max?: number;
  icon?: () => JSX.Element;
  showInput?: boolean;
  isDecimal?: boolean;
}

export interface InputControllerProps extends Partial<CommonClassesInterface> {
  type: 'number' | 'text' | 'email' | 'tel' | 'range' | 'password' | 'date';
  key?: string;
  name: string;
  label?: string;
  placeholder?: string;
  isRequired?: boolean;
  defaultValue?: string;
  isDisabled?: boolean;
  isReadOnly?: boolean;
  min?: number;
  max?: number;
  icon?: () => JSX.Element;
  isFractional?: boolean;
  events?: React.InputHTMLAttributes<HTMLInputElement>;
}

export interface DropdownControllerProps
  extends Partial<CommonClassesInterface> {
  type: 'dropdown';
  key?: string;
  name: string;
  label?: string;
  placeholder?: string;
  options?: string[] | number[];
  isId?: boolean;
  isRequired?: boolean;
  defaultValue?: string;
  isDisabled?: boolean;
  icon?: () => JSX.Element;
  onChange?: (inputValue: string) => void;
  onSelected?: (selectedValue: string) => void;
  hideLabel?: boolean;
  disabled?: boolean;
}

export interface DateControllerProps extends Partial<CommonClassesInterface> {
  type: 'date';
  name: string;
  label?: string;
  placeholder?: string;
  isRequired?: boolean;
  defaultValue?: string;
  isDisabled?: boolean;
  min?: string | Date;
  max?: string | Date;
  excludeDateIntervals?: { start: Date; end: Date }[];
  icon?: () => JSX.Element;
  filterDates?: (date: undefined) => boolean;
}

export interface TextAreaControllerProps
  extends Partial<CommonClassesInterface> {
  type: 'textarea';
  name: string;
  label?: string;
  placeholder?: string;
  rows?: number;
  multiline?: number;
  isRequired?: boolean;
  defaultValue?: string;
  isDisabled?: boolean;
  icon?: () => JSX.Element;
}

export interface fileControllerProps extends Partial<CommonClassesInterface> {
  type: 'file';
  name: string;
  label?: string;
  memTypes?: string;
  placeholder?: string;
  rows?: number;
  key?: string;
  isMultiple?: boolean;
  isRequired?: boolean;
  defaultValue?: string;
  isDisabled?: boolean;
  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
  icon?: () => JSX.Element;
}

export interface CheckBoxControllerProps
  extends Partial<CommonClassesInterface> {
  type: 'checkbox';
  name: string;
  label?: string;
  placeholder?: string;
  isRequired?: boolean;
  defaultValue?: boolean;
  isDisabled?: boolean;
  icon?: () => JSX.Element;
  onChange?: Dispatch<SetStateAction<boolean>>;
}

export interface SubscriptionCreatePropType {
  setShowCreateView?: Dispatch<SetStateAction<boolean>>;
}

export interface RadioButtonControllerProps
  extends Partial<CommonClassesInterface> {
  type: 'radioButton';
  name: string;
  label?: string;
  placeholder?: string;
  isRequired?: boolean;
  isDisabled?: boolean;
  options?: string[] | number[];
  icon?: () => JSX.Element;
}

export interface MultiCheckBoxControllerProps
  extends Partial<CommonClassesInterface> {
  type: 'multiCheckbox';
  name: string;
  label?: string;
  options: { key: string; label: string | number }[];
  placeholder?: string;
  isRequired?: boolean;
  defaultValue?: string;
  isDisabled?: boolean;
  icon?: () => JSX.Element;
}